Convert back to bit-banging
Ended up going back to bit-banging.
The display need to be pushed 36 bits for a single update. The SPI library is only able to push whole bytes. This resulted in unreliable updates to the display and one would see garbled text quite often.
The code actually ends up a bit cleaner this way (in my opinion).